feat: 点击弹幕、昵称、礼物后选项功能
